extract DelegateStore
